+++ src/components/index.ts
@@ -98,6 +98,11 @@ import InputCourse from './inputDropdown/inputCourse/InputCourse';
 import InputProfessor from './inputDropdown/inputCourse/InputProfessor';
 import InputRole from './inputDropdown/inputRole/InputRole';
 import InputRoleForm from './inputDropdown/inputRole/InputRoleForm';
+import Description from './recruit/create/Description';
+import BasicInformation from './recruit/create/BasicInformation';
+import DetailedInformation from './recruit/create/DetailedInformation';
+import RecruitRoles from './recruit/create/RecruitRoles';
+import ControlButtons from './recruit/create/ControlButtons';
 
 export {
 	Header,
@@ -191,4 +196,9 @@ export {
 	InputProfessor,
 	InputRole,
 	InputRoleForm,
+	Description,
+	BasicInformation,
+	DetailedInformation,
+	RecruitRoles,
+	ControlButtons,
 };
