+++ src/components/recruit/create/ControlButtons.tsx
@@ -1,0 +1,16 @@
+import React from 'react';
+
+const ControlButtons = () => {
+	return (
+		<section className='container-btns'>
+			<button type='button' className='cancel-btn'>
+				취소
+			</button>
+			<button type='submit' className='submit-btn'>
+				등록
+			</button>
+		</section>
+	);
+};
+
+export default ControlButtons;
