+++ src/components/recruit/create/Description.tsx
@@ -1,0 +1,18 @@
+import React from 'react';
+
+const Description = () => {
+	return (
+		<>
+			<section className='container-introduction'>
+				<h2>함께할 멤버들에게 알려줄 정보들을 작성해주세요.</h2>
+				<h4>
+					작성하신 구인글 정보들은 구인게시판을 통해 미래의 멤버들에게 보여집니다. 알려주고 싶은
+					정보들을 자유롭게 작성해보세요!
+				</h4>
+			</section>
+			<hr className='under-intro' />
+		</>
+	);
+};
+
+export default Description;
